Resolve a bunch of item references

+++ src/main/java/net/errorcraft/itematic/access/entity/EntityAccess.java
@@ -7,15 +7,15 @@ import net.minecraft.registry.entry.RegistryEntry;
 
 public interface EntityAccess {
     default ItemEntity itematic$dropItem(RegistryKey<Item> key) {
-        return this.itematic$dropItem(key, 0);
+        return this.itematic$dropItem(key, 0.0f);
     }
-    default ItemEntity itematic$dropItem(RegistryKey<Item> key, int yOffset) {
+    default ItemEntity itematic$dropItem(RegistryKey<Item> key, float yOffset) {
         return null;
     }
     default ItemEntity itematic$dropItem(RegistryEntry<Item> entry) {
-        return this.itematic$dropItem(entry, 0);
+        return this.itematic$dropItem(entry, 0.0f);
     }
-    default ItemEntity itematic$dropItem(RegistryEntry<Item> entry, int yOffset) {
+    default ItemEntity itematic$dropItem(RegistryEntry<Item> entry, float yOffset) {
         return null;
     }
 }
